edge transformation changes to match 0.1.1
