+++ src/main/groovy/com/github/eugene/kamenev/orient/ast/util/EntityProperty.groovy
@@ -50,8 +50,6 @@ abstract class EntityProperty {
         ASTUtil.eachMethodArg(methodCall, closure)
     }
 
-
-
     /**
      * if property is mapped as collection
      *
