+++ src/main/groovy/com/github/eugene/kamenev/orient/document/OrientDocumentTransformation.groovy
@@ -13,6 +13,7 @@ import org.codehaus.groovy.control.SourceUnit
 import org.codehaus.groovy.transform.AbstractASTTransformation
 import org.codehaus.groovy.transform.DelegateASTTransformation
 import org.codehaus.groovy.transform.GroovyASTTransformation
+
 /**
  * Transformation for OrientDB document usage
  * @since 0.1.0
